refactor: Handle concurrency issues in Plot3dTab, unmount handlers

+++ src/colorizer/types.ts
@@ -380,5 +380,5 @@ export type VectorFieldData = {
   xData: Float32Array;
   yData: Float32Array;
   zData: Float32Array;
-  count: Uint16Array;
+  count: Uint32Array;
 };
